pc-server/requirements.txt
customtkinter>=5.2.0
pyaudio>=0.2.13
numpy>=1.24.0
pystray>=0.19.0
Pillow>=10.0.0
sounddevice>=0.4.6
opuslib>=3.0.1
pyinstaller>=6.0.0
